Create main communication loop

+++ src/main/java/model/messageparams/IntersectionResponse.java
@@ -4,9 +4,9 @@ import java.io.Serializable;
 
 public class IntersectionResponse implements Serializable {
     public final Float speed;
-    public final long time;
+    public final Float time;
 
-    public IntersectionResponse(Float speed, long time) {
+    public IntersectionResponse(Float speed, Float time) {
         this.speed = speed;
         this.time = time;
     }
